Makes vim brighter
Used a dark background for a long time, but a dark setup is hard top use
on bright environments (e.g: coding on Ipiranga Museum).

Just for future notices, Nelson started on dark background and is loving
it. So probably it is just a matter of getting too used to a color
scheme just to get bored to death by it.