Specifies file type plugins for shell and PHP
Rather than having them in global options, defines better defaults for
each one of them (as they are the ones I usually write most).

For PHP, PHPClassExplorer is moved as well as the "trailing white space
remover" befor the buffer is written. Adds the map `<leader>l` to lint a
PHP file.

For shell, `noexpandtab` is added as I am a huge fan of the `<<-`
heredocs which only work on tabs.