Prefer The Silver Searcher on VIM
It is way faster for everything, uses it as the grep program and the
CTRL-P searcher as well. After a week I realized that CTRL-P does not
even depend on cache anymore, which is awesome.

To lose the muscle memory of doing `:vimgrep` every time, I mapped
`<leader>s` to `:grep` the work the cursor is on and open the QuickFix
list. I seriously need some mappings for the quickfix now, mapping the
arrow keys would be awesome.

[1]: https://github.com/ggreer/the_silver_searcher