Moves aliases to .bash_profile
Aliases are a mess, using multiple operating systems and my current
organization is making me unease currently. Maybe it is time to rethink
a litle bit of this and make something like:

	for s in `ls "$HOME/.warwick-*"`
	do
		echo "  Loading $s"
		source "$s"
	done

I like how I bootstrap everything for OSX in the past and how I am
missing this on Linux, specially when I am trying out Windows Susytem
for Linux and hopping into servers.