Proper WSL home directory detection
Fixes a bug when using WSL from an editor (ex: VSCode) where the
directory of WSL was not the user home but the project root directory.
Previous to this commit, Git configuration would get symlinked to
the project. Now the home directory of the user, on Windows, is properly
defined without relying on `pwd` for that.