Interactive remote branch checkout with "gcr"
Before changing branches, stash changes to allow the command to run.
Usually this is what is always wanted, in this way work is not lost.
Works for all "checkouts", not only the one introduced in this change.

Staged changes are stashed separately to allow better control when
popping them, trying to minimize the amount of commands needed to run to
restore work.

Introduces remote checkout of branches as well - interactive.