Integrates git (ssh) with 1password
Private keys are stored inside 1password, with ssh-agent, it is
possible to negotiate keys with SSH without moving them on
disk.

They work differently on Windows and MacOS though. On windows, which is
where this commit is being tested, we need to use `ssh.exe` in order to
interact with 1password. As `.gitconfig` needs to be configured to use
the right binary `ssh-win-or-not` is introduced - great name isn't it?